+++ tools/buildInstaller.js
@@ -47,6 +47,7 @@ if (isDarwin) {
      // Create an update zip
     'ditto -c -k --sequesterRsrc --keepParent ' + buildDir + '/Brave.app dist/Brave-' + VersionInfo.braveVersion + '.zip'
   ]
+  execute(cmds, {}, console.log.bind(null, 'done'))
 } else if (isWindows) {
   // a cert file must be present to sign the created package
   // a password MUST be passed as the CERT_PASSWORD environment variable
@@ -60,11 +61,28 @@ if (isDarwin) {
   // need to store the output files in separate directories
   outDir = path.join(outDir, arch)
 
-  cmds = [
-    'electron-installer-squirrel-windows "' + buildDir + '" --platform=win --out="' + outDir + '" --name=brave --product_name="Brave" --overwrite --debug --loading_gif="res/brave_splash_installing.gif" --setup_icon=res/app.ico --sign_with_params=' + format('"-a -fd sha256 -f \\"%s\\" -p \\"%s\\" -t http://timestamp.verisign.com/scripts/timstamp.dll"', path.resolve(cert), certPassword),
-    `mv ${outDir}/BraveSetup.exe ${outDir}/BraveSetup-${arch}.exe`,
-    `mv ${outDir}/Setup.msi ${outDir}/BraveSetup-${arch}.msi`
-  ]
+  var electronInstaller = require('electron-winstaller')
+  var resultPromise = electronInstaller.createWindowsInstaller({
+    appDirectory: buildDir,
+    outputDirectory: outDir,
+    title: 'Brave',
+    authors: 'Brave Software',
+    loadingGif: 'res/brave_splash_installing.gif',
+    setupIcon: 'res/app.ico',
+    iconUrl: 'https://brave.com/favicon.ico',
+    signWithParams: format('-a -fd sha256 -f "%s" -p "%s" -t http://timestamp.verisign.com/scripts/timstamp.dll', path.resolve(cert), certPassword),
+    exe: 'Brave.exe'
+  })
+  resultPromise.then(() => {
+    cmds = [
+      `mv ${outDir}/Setup.exe ${outDir}/BraveSetup-${arch}.exe`,
+      `mv ${outDir}/Setup.msi ${outDir}/BraveSetup-${arch}.msi`,
+      // Change the casing of the output file name, Windows won't allow this in a single mv
+      `mv ${outDir}/brave-${VersionInfo.braveVersion}-full.nupkg ${outDir}/Brave-${VersionInfo.braveVersion}-full.nupkg2`,
+      `mv ${outDir}/Brave-${VersionInfo.braveVersion}-full.nupkg2 ${outDir}/Brave-${VersionInfo.braveVersion}-full.nupkg`
+    ]
+    execute(cmds, {}, console.log.bind(null, 'done'))
+  }, (e) => console.log(`No dice: ${e.message}`))
 } else if (isLinux) {
   console.log('Install with sudo dpkg -i dist/brave_' + VersionInfo.braveVersion + '_amd64.deb')
   console.log('Or install with sudo rpm -i dist/brave_' + VersionInfo.braveVersion + '.amd64.rpm')
@@ -84,9 +102,9 @@ if (isDarwin) {
     // .tar.bz2 file
     'tar -jcvf dist/Brave.tar.bz2 ./Brave-linux-x64'
   ]
+  execute(cmds, {}, console.log.bind(null, 'done'))
 } else {
   console.log('Installer not supported for platform: ' + process.platform)
   process.exit(1)
 }
 
-execute(cmds, {}, console.log.bind(null, 'done'))